+++ python/sglang/multimodal_gen/runtime/disaggregation/transport/codec.py
