+++ python/sglang/multimodal_gen/runtime/disaggregation/transport/manager.py
@@ -7,11 +7,11 @@ from dataclasses import dataclass, field
 
 import torch
 
-from sglang.multimodal_gen.runtime.disaggregation.transport.rdma.transfer_buffer import (
+from sglang.multimodal_gen.runtime.disaggregation.transport.buffer import (
     SlotHandle,
     TransferTensorBuffer,
 )
-from sglang.multimodal_gen.runtime.disaggregation.transport.rdma.transfer_engine import (
+from sglang.multimodal_gen.runtime.disaggregation.transport.engine import (
     BaseTransferEngine,
 )
