--- python/sglang/multimodal_gen/runtime/disaggregation/transport/rdma/__init__.py
@@ -1,2 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-"""RDMA transport: tensor transfer via Mooncake TransferEngine."""
