+++ python/sglang/multimodal_gen/test/unit/test_transfer_allocator.py
@@ -4,7 +4,7 @@
 import threading
 import unittest
 
-from sglang.multimodal_gen.runtime.disaggregation.transport.rdma.transfer_allocator import (
+from sglang.multimodal_gen.runtime.disaggregation.transport.allocator import (
     BuddyAllocator,
 )
