+++ python/sglang/multimodal_gen/test/unit/test_transfer_engine.py
@@ -4,7 +4,7 @@
 import ctypes
 import unittest
 
-from sglang.multimodal_gen.runtime.disaggregation.transport.rdma.transfer_engine import (
+from sglang.multimodal_gen.runtime.disaggregation.transport.engine import (
     BaseTransferEngine,
     MockTransferEngine,
     create_transfer_engine,
