+++ python/sglang/multimodal_gen/test/unit/test_transfer_manager.py
@@ -5,18 +5,18 @@ import unittest
 
 import torch
 
-from sglang.multimodal_gen.runtime.disaggregation.transport.rdma.transfer_buffer import (
+from sglang.multimodal_gen.runtime.disaggregation.transport.buffer import (
     TransferTensorBuffer,
 )
-from sglang.multimodal_gen.runtime.disaggregation.transport.rdma.transfer_engine import (
+from sglang.multimodal_gen.runtime.disaggregation.transport.engine import (
     MockTransferEngine,
 )
-from sglang.multimodal_gen.runtime.disaggregation.transport.rdma.transfer_manager import (
+from sglang.multimodal_gen.runtime.disaggregation.transport.manager import (
     DiffusionTransferManager,
     PendingReceive,
     StagedTransfer,
 )
-from sglang.multimodal_gen.runtime.disaggregation.transport.transfer_protocol import (
+from sglang.multimodal_gen.runtime.disaggregation.transport.protocol import (
     TransferAllocMsg,
     TransferPushedMsg,
     TransferPushMsg,
